% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{adult}
\alias{adult}
\title{Adult Dataset}
\format{
Um data frame com 48842 linhas e 15 variáveis:
\describe{
\item{age}{}
\item{workclass}{}
\item{fnlwgt}{}
\item{education}{}
\item{education_num}{}
\item{marital_status}{}
\item{occupation}{}
\item{relationship}{}
\item{race}{}
\item{sex}{}
\item{capital_gain}{}
\item{capital_loss}{}
\item{hours_per_week}{}
\item{native_country}{}
\item{more_than_50k}{variável resposta com duas categorias: "yes" e "no"}
}
}
\source{
\url{http://archive.ics.uci.edu/ml/datasets/Adult}
}
\usage{
adult
}
\description{
Extração feita por Barry Becker do Censo de 1994 dos Estados Unidos.
}
\details{
Objetivo de predição é determinar se a pessoa ganha acima de 50 mil dólares por ano.
}
\keyword{datasets}
