% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_base.R
\name{bases_disponiveis}
\alias{bases_disponiveis}
\title{See available datasets in the package}
\usage{
bases_disponiveis()
}
\value{
Characeter vector with dataset names.
}
\description{
See available datasets in the package
}
\examples{
bases_disponiveis()
}
\seealso{
\code{\link[=pegar_base]{pegar_base()}} to import a dataset.
}
