% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{covid}
\alias{covid}
\title{Painel Coronavírus no Brasil}
\format{
Um data frame com 383908 linhas e 19 variáveis:
\describe{
\item{regiao}{}
\item{estado}{}
\item{municipio}{}
\item{coduf}{}
\item{codmun}{}
\item{codRegiaoSaude}{}
\item{nomeRegiaoSaude}{}
\item{data}{}
\item{semanaEpi}{}
\item{populacaoTCU2019}{}
\item{casosAcumulado}{}
\item{casosNovos}{}
\item{obitosAcumulado}{}
\item{obitosNovos}{}
\item{Recuperadosnovos}{}
\item{emAcompanhamentoNovos}{}
\item{eh_capital}{}
\item{obitosAcumulado_log2}{}
\item{obitosNovos_log2}{}
}
}
\source{
\url{https://covid.saude.gov.br/}
}
\usage{
covid
}
\description{
Contagem de casos confirmados e óbitos confirmados por dia no Brasil divulgado pelo portal \url{https://covid.saude.gov.br/}.
}
\keyword{datasets}
