% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pegar_base.R
\name{pegar_base}
\alias{pegar_base}
\title{Import a dataset}
\usage{
pegar_base(base)
}
\arguments{
\item{base}{String with the dataset name you want to import.
See the \code{\link[=bases_disponiveis]{bases_disponiveis()}} function.}
}
\value{
A dataset.
}
\description{
Import a dataset
}
\examples{
pegar_base("imdb")
}
\seealso{
\code{\link[=bases_disponiveis]{bases_disponiveis()}} for available datasets.
}
